add faq sections for detailed TVL
